# docs/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.SkillItem:
    properties:
      content:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  handlers.SkillsDeleteRequest:
    properties:
      names:
        items:
          type: string
        type: array
    type: object
  handlers.SkillsResponse:
    properties:
      skills:
        items:
          $ref: '#/definitions/handlers.SkillItem'
        type: array
    type: object
  handlers.SkillsUpsertRequest:
    properties:
      skills:
        items:
          $ref: '#/definitions/handlers.SkillItem'
        type: array
    type: object
...
  handlers.skillsOpResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  subagent.AddSkillsRequest:
    properties:
      skills:
        items:
          type: string
        type: array
    type: object
...
  subagent.SkillsResponse:
    properties:
      skills:
        items:
          type: string
        type: array
    type: object
...
  subagent.UpdateSkillsRequest:
    properties:
      skills:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /container/skills:
    delete:
      parameters:
      - description: Delete skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SkillsDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.skillsOpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete skills from container
      tags:
      - containerd
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SkillsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List skills from container
      tags:
      - containerd
    post:
      parameters:
      - description: Skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SkillsUpsertRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.skillsOpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Upload skills into container
      tags:
      - containerd
...
  /subagents/{id}/skills:
    get:
      description: Get a subagent's skills
      parameters:
      - description: Subagent ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subagent.SkillsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get subagent skills
      tags:
      - subagent
    post:
      description: Add skills to a subagent
      parameters:
      - description: Subagent ID
        in: path
        name: id
        required: true
        type: string
      - description: Skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/subagent.AddSkillsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subagent.SkillsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Add subagent skills
      tags:
      - subagent
    put:
      description: Replace a subagent's skills
      parameters:
      - description: Subagent ID
        in: path
        name: id
        required: true
        type: string
      - description: Skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/subagent.UpdateSkillsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subagent.SkillsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update subagent skills
      tags:
      - subagent
swagger: "2.0"